+++ CHANGELOG.md
@@ -22,6 +22,10 @@
 * Added: Accumulo additional client configuration options (config parameter `clientConfiguration.xxx=<value>`)
 * Changed: Deprecated Elasticsearch global scoring strategy  
 
+# v4.1.1.2
+* Added: Query edges with a given in or out vertex
+* Added: Elasticsearch: Return ElasticsearchEdge and ElasticsearchVertex when FetchHints.EDGE_REF and FetchHints.NONE is set for performance
+
 # v4.1.1.1
 * Fixed: `graph.findRelatedEdgeSummary` filtering edges that were re-added after soft delete/hidden